Refactor ComponentSettings
